split confirmation schemas
